refactor(session-persistence): extract reconciliation owner
