+++ scripts/ci/module-impact.json
@@ -202,6 +202,8 @@
       "ownerPaths": [
         "src/main/session-persistence/coordinator.ts",
         "src/main/session-persistence/deletion-owner.ts",
+        "src/main/session-persistence/legacy-upload.ts",
+        "src/main/session-persistence/reconciliation-owner.ts",
         "src/main/session-persistence/state-owner.ts"
       ],
       "interfacePaths": ["src/main/session-persistence/coordinator.ts"],
