+++ src/main/session-persistence/legacy-upload.ts
@@ -1,0 +1,8 @@
+import type { PersistedChatSession } from '../../shared/session-persistence'
+
+const hasLegacySessionUpload = (session: PersistedChatSession): boolean =>
+  [...session.messages, ...(session.conversationGraph?.messages ?? [])].some((message) =>
+    message.uploads?.some((upload) => !upload.versionId)
+  )
+
+export { hasLegacySessionUpload }
